removing dead code for .zip file manipulation
Also moved the 'decodeBase64Image' function out of the completely unrelated zip-utils.js
