+++ package.json
@@ -40,7 +40,6 @@
     "codecov": "npm run test && (codecov || true)"
   },
   "dependencies": {
-    "archiver": "^1.0.0",
     "async": "^2.0.0-rc.5",
     "body-parser": "^1.15.1",
     "cf-deployment-tracker-client": "0.0.8",
